Moved debug (backwards incomptible, bug debug isn't part of core stable API)
