--- src/main/scala/parsley/Breakpoint.scala
@@ -1,7 +1,0 @@
-package parsley
-
-trait Breakpoint
-case object NoBreak extends Breakpoint
-case object EntryBreak extends Breakpoint
-case object ExitBreak extends Breakpoint
-case object FullBreak extends Breakpoint
