+++ src/main/scala/parsley/Parsley.scala
@@ -224,14 +224,6 @@ object Parsley
           */
         def unexpected(msggen: A => String): Parsley[Nothing] = new Parsley(new deepembedding.FastUnexpected(p.internal, msggen))
         /**
-          * Using this method enables debugging functionality for this parser. When it is entered a snapshot is taken and
-          * presented on exit. It will signify when a parser is entered and exited as well. Use the break parameter to halt
-          * execution on either entry, exit, both or neither.
-          * @param name The name to be assigned to this parser
-          * @param break The breakpoint properties of this parser, defaults to NoBreak
-          */
-        def debug[A_ >: A](name: String, break: Breakpoint = NoBreak): Parsley[A_] = new Parsley(new deepembedding.Debug[A_](p.internal, name, break))
-        /**
           * A fold for a parser: `p.foldRight(k)(f)` will try executing `p` many times until it fails, combining the
           * results with right-associative application of `f` with a `k` at the right-most position
           *
